Rename header related files to kebab case
